+++ backends/sql/metadatafunctions.php
@@ -3,7 +3,7 @@
 class romprmetadata {
 
 	public static function sanitise_data(&$data) {
-	
+
 		foreach (array( 'action',
 						'title',
 						'artist',
@@ -56,7 +56,7 @@ class romprmetadata {
 		}
 	}
 
-	public static function set($data) {
+	public static function set($data, $keep_wishlist = false) {
 		global $returninfo;
 		if ($data['artist'] === null ||
 			$data['title'] === null ||
@@ -83,7 +83,7 @@ class romprmetadata {
 
 		$newttids = array();
 		foreach ($ttids as $ttid) {
-			if (!track_is_wishlist($ttid)) {
+			if ($keep_wishlist || !track_is_wishlist($ttid)) {
 				$newttids[] = $ttid;
 			}
 		}
@@ -111,12 +111,12 @@ class romprmetadata {
 		// This is used for adding specific tracks so we need urionly to be true
 		// We don't simply call into this using 'set' with urionly set to true
 		// because that might result in the rating being changed
-		
+
 		// The only time we call inot this with $urionly set to false is when we're restoring a metadata
 		// backup. In that case we might be copying data from one setup to another and we might have
 		// the track already in local, so we don't want to add duplicates. Neither way is perfect but
 		// this makes most sense I think.
-		
+
 		global $returninfo;
 		$ttids = romprmetadata::find_item($data, $urionly);